Single receive in bus linker (#2359)
- separate the native and bus linker as two implementations of a
`LinkerBackend` trait
- refactor the object data structures to simplify linker implementation
- change the bus linker to use the bus directly and use a single receive
per operation, with the `interaction_id` being hash based
- pull the permutation selector array treatment into the linker, since
it should be linker specific (per link for native, per operation for
bus)

---------
