+++ linker/Cargo.toml
@@ -18,6 +18,8 @@ pretty_assertions = "1.4.0"
 itertools = "0.13"
 lazy_static = "1.4.0"
 log = "0.4.22"
+sha2 = { version = "0.10.8", default-features = false }
+ibig = { version = "0.3.6", default-features = false }
 
 [dev-dependencies]
 powdr-airgen.workspace = true
