+++ test_data/std/bus_lookup.asm
@@ -1,4 +1,4 @@
-use std::protocols::lookup_via_bus::lookup_send;
+use std::protocols::bus::bus_send;
 use std::protocols::lookup_via_bus::lookup_receive;
 
 machine Main with degree: 8 {
@@ -11,7 +11,6 @@ machine Main with degree: 8 {
     col fixed INC_Y = [2, 3, 4, 5, 6, 7, 8, 9];
 
     let LOOKUP_ID = 42;
-    let lookup_constraint = [x, y] in [INC_X, INC_Y];
-    lookup_send(LOOKUP_ID, lookup_constraint);
-    lookup_receive(LOOKUP_ID, lookup_constraint, 1);
+    bus_send(LOOKUP_ID, [x, y], 1);
+    lookup_receive(LOOKUP_ID, 1, [INC_X, INC_Y]);
 }
